Move MVC Config to ServletConfig instead of root Application
