+++ airsonic-main/src/main/java/org/airsonic/player/Application.java
@@ -1,7 +1,5 @@
 package org.airsonic.player;
 
-import org.airsonic.player.filter.*;
-import org.directwebremoting.servlet.DwrServlet;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -10,13 +8,8 @@ import org.springframework.boot.autoconfigure.web.MultipartAutoConfiguration;
 import org.springframework.boot.builder.SpringApplicationBuilder;
 import org.springframework.boot.context.embedded.ConfigurableEmbeddedServletContainer;
 import org.springframework.boot.context.embedded.EmbeddedServletContainerCustomizer;
-import org.springframework.boot.web.servlet.FilterRegistrationBean;
-import org.springframework.boot.web.servlet.ServletRegistrationBean;
 import org.springframework.boot.web.support.SpringBootServletInitializer;
-import org.springframework.context.annotation.Bean;
 import org.springframework.util.ReflectionUtils;
-
-import javax.servlet.Filter;
 
 import java.lang.reflect.Method;
 
@@ -27,132 +20,6 @@ import java.lang.reflect.Method;
 public class Application extends SpringBootServletInitializer implements EmbeddedServletContainerCustomizer {
 
     private static final Logger LOG = LoggerFactory.getLogger(Application.class);
-
-    /**
-     * Registers the DWR servlet.
-     *
-     * @return a registration bean.
-     */
-    @Bean
-    public ServletRegistrationBean dwrServletRegistrationBean() {
-        ServletRegistrationBean servlet = new ServletRegistrationBean(new DwrServlet(), "/dwr/*");
-        servlet.addInitParameter("crossDomainSessionSecurity","false");
-        return servlet;
-    }
-
-    @Bean
-    public ServletRegistrationBean cxfServletBean() {
-        return new ServletRegistrationBean(new org.apache.cxf.transport.servlet.CXFServlet(), "/ws/*");
-    }
-
-    @Bean
-    public FilterRegistrationBean bootstrapVerificationFilterRegistration() {
-        FilterRegistrationBean registration = new FilterRegistrationBean();
-        registration.setFilter(bootstrapVerificationFiler());
-        registration.addUrlPatterns("/*");
-        registration.setName("BootstrapVerificationFilter");
-        registration.setOrder(1);
-        return registration;
-    }
-
-    @Bean
-    public Filter bootstrapVerificationFiler() {
-        return new BootstrapVerificationFilter();
-    }
-
-    @Bean
-    public FilterRegistrationBean parameterDecodingFilterRegistration() {
-        FilterRegistrationBean registration = new FilterRegistrationBean();
-        registration.setFilter(parameterDecodingFilter());
-        registration.addUrlPatterns("/*");
-        registration.setName("ParameterDecodingFilter");
-        registration.setOrder(2);
-        return registration;
-    }
-
-    @Bean
-    public Filter parameterDecodingFilter() {
-        return new ParameterDecodingFilter();
-    }
-
-    @Bean
-    public FilterRegistrationBean restFilterRegistration() {
-        FilterRegistrationBean registration = new FilterRegistrationBean();
-        registration.setFilter(restFilter());
-        registration.addUrlPatterns("/rest/*");
-        registration.setName("RESTFilter");
-        registration.setOrder(3);
-        return registration;
-    }
-
-    @Bean
-    public Filter restFilter() {
-        return new RESTFilter();
-    }
-
-    @Bean
-    public FilterRegistrationBean requestEncodingFilterRegistration() {
-        FilterRegistrationBean registration = new FilterRegistrationBean();
-        registration.setFilter(requestEncodingFilter());
-        registration.addUrlPatterns("/*");
-        registration.addInitParameter("encoding", "UTF-8");
-        registration.setName("RequestEncodingFilter");
-        registration.setOrder(4);
-        return registration;
-    }
-
-    @Bean
-    public Filter requestEncodingFilter() {
-        return new RequestEncodingFilter();
-    }
-
-    @Bean
-    public FilterRegistrationBean cacheFilterRegistration() {
-        FilterRegistrationBean registration = new FilterRegistrationBean();
-        registration.setFilter(cacheFilter());
-        registration.addUrlPatterns("/icons/*", "/style/*", "/script/*", "/dwr/*", "/icons/*", "/coverArt.view", "/avatar.view");
-        registration.addInitParameter("Cache-Control", "max-age=36000");
-        registration.setName("CacheFilter");
-        registration.setOrder(5);
-        return registration;
-    }
-
-    @Bean
-    public Filter cacheFilter() {
-        return new ResponseHeaderFilter();
-    }
-
-    @Bean
-    public FilterRegistrationBean noCacheFilterRegistration() {
-        FilterRegistrationBean registration = new FilterRegistrationBean();
-        registration.setFilter(noCacheFilter());
-        registration.addUrlPatterns("/statusChart.view", "/userChart.view", "/playQueue.view", "/podcastChannels.view", "/podcastChannel.view", "/help.view", "/top.view", "/home.view");
-        registration.addInitParameter("Cache-Control", "no-cache, post-check=0, pre-check=0");
-        registration.addInitParameter("Pragma", "no-cache");
-        registration.addInitParameter("Expires", "Thu, 01 Dec 1994 16:00:00 GMT");
-        registration.setName("NoCacheFilter");
-        registration.setOrder(6);
-        return registration;
-    }
-
-    @Bean
-    public Filter metricsFilter() {
-        return new MetricsFilter();
-    }
-
-    @Bean
-    public FilterRegistrationBean metricsFilterRegistration() {
-        FilterRegistrationBean registration = new FilterRegistrationBean();
-        registration.setFilter(metricsFilter());
-        registration.setOrder(7);
-        return registration;
-    }
-
-
-    @Bean
-    public Filter noCacheFilter() {
-        return new ResponseHeaderFilter();
-    }
 
     private static SpringApplicationBuilder doConfigure(SpringApplicationBuilder application) {
         // Customize the application or call application.sources(...) to add sources
